Extract processors to separate files
