+++ src/controllers/processors/UserEmailFormSubmissionProcessor.ts
@@ -1,0 +1,47 @@
+import { Maybe } from 'ch-node-session-handler';
+import { SessionKey } from 'ch-node-session-handler/lib/session/keys/SessionKey';
+import { SignInInfoKeys } from 'ch-node-session-handler/lib/session/keys/SignInInfoKeys';
+import { ISignInInfo, IUserProfile } from 'ch-node-session-handler/lib/session/model/SessionInterfaces';
+import { Request } from 'express';
+import { inject } from 'inversify';
+import { provide } from 'inversify-binding-decorators';
+
+import { FormSubmissionProcessor } from 'app/controllers/processors/FormSubmissionProcessor';
+import { Appeal, APPEALS_KEY } from 'app/models/Appeal';
+import { Email } from 'app/modules/email-publisher/Email';
+import { EmailService } from 'app/modules/email-publisher/EmailService';
+
+@provide(UserEmailFormSubmissionProcessor)
+export class UserEmailFormSubmissionProcessor implements FormSubmissionProcessor {
+    constructor(@inject(EmailService) private readonly emailService: EmailService) {}
+
+    async process(req: Request): Promise<void> {
+        const userProfile = req.session
+            .chain(_ => _.getValue<ISignInInfo>(SessionKey.SignInInfo))
+            .map(info => info[SignInInfoKeys.UserProfile])
+            .orDefault({});
+
+        const appealsData = req.session
+            .chain(_ => _.getExtraData())
+            .chain(data => Maybe.fromNullable(data[APPEALS_KEY]))
+            .orDefault({});
+
+        await this.emailService.send(this.buildEmail(userProfile as IUserProfile, appealsData));
+    }
+
+    private buildEmail(userProfile: IUserProfile, appeal: Appeal): Email {
+        return {
+            to: userProfile?.email as string,
+            subject: `Confirmation of your appeal - ${appeal.penaltyIdentifier.companyNumber} - Companies House`,
+            body: {
+                templateName: 'lfp-appeal-submission-confirmation',
+                templateData: {
+                    companyNumber: appeal.penaltyIdentifier.companyNumber,
+                    userProfile: {
+                        email: userProfile?.email
+                    }
+                }
+            }
+        };
+    }
+}
